Fetch data from discovery server for getListings and getListing calls
